fix(review): add post-negation check to blocking result signals (#2002)
Address gemini review findings:
1. summary_line_has_blocking_result_signal now checks for negation words
   after signal tokens ("found none", "reported previously") to prevent
   false-positive FAILs.
2. refresh_and_repair_result_from_dir uses atomic temp+rename instead of
   truncating fs::write for cross-project result repair.
3. Split oversized test file: #1978/#2002 tests moved to
   session_cmds_tests_tail_wait_2002.rs.
